solution_intelligence/requirements.txt
requests
sentence-transformers
hdbscan
numpy
scikit-learn
tqdm
faiss-cpu
